# docs/swagger.yaml
basePath: /
definitions:
  models.Admin:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      password:
        type: string
      restaurant:
        $ref: '#/definitions/models.Restaurant'
    required:
    - email
    - name
    - password
    type: object
  models.Dish:
    properties:
      category:
        type: string
      currency:
        type: string
      description:
        type: string
      id:
        type: string
      ingredients:
        items:
          type: string
        type: array
      mass_grams:
        type: integer
      menu_id:
        type: string
      name:
        type: string
      price:
        type: integer
      type:
        type: string
    required:
    - category
    - currency
    - description
    - ingredients
    - mass_grams
    - name
    - price
    - type
    type: object
...
  models.Menu:
    properties:
      description:
        type: string
      dish:
        items:
          $ref: '#/definitions/models.Dish'
        type: array
      id:
        type: string
      name:
        type: string
      qrcode:
        items:
          type: integer
        type: array
      qrcodeID:
        type: string
      restaurant_id:
        type: string
    required:
    - description
    - name
    type: object
  models.apimodels.Message:
    properties:
      apimodels.Message:
        type: string
    type: object
  models.ReservAPI:
    properties:
      day:
        type: integer
      hour:
        type: integer
      id:
        type: string
      minute:
        type: integer
      month:
        type: integer
      reserver_name:
        type: string
      reserver_phone:
        type: string
      second:
        type: integer
      table_id:
        type: string
      year:
        type: integer
    required:
    - day
    - hour
    - minute
    - month
    - reserver_name
    - reserver_phone
    - table_id
    - year
    type: object
  models.ReservDB:
    properties:
      id:
        type: string
      reservation_time:
        type: string
      reserver_name:
        type: string
      reserver_phone:
        type: string
      restaurant_id:
        type: string
      table_id:
        type: string
    type: object
  models.Restaurant:
    properties:
      admin_id:
        type: string
      id:
        type: string
      location:
        type: string
      menu:
        $ref: '#/definitions/models.Menu'
      name:
        type: string
      staff:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
      table:
        items:
          $ref: '#/definitions/models.Table'
        type: array
    required:
    - location
    - name
    type: object
  models.Staff:
    properties:
      age:
        type: integer
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
      restaurant_id:
        type: string
    required:
    - age
    - email
    - gender
    - name
    - phone
    - position
    type: object
  models.Table:
    properties:
      id:
        type: string
      is_occupied:
        type: boolean
      is_reserved:
        type: boolean
      max_people:
        type: integer
      number:
        type: integer
      placement:
        type: string
      reservation:
        $ref: '#/definitions/models.ReservDB'
      restaurant_id:
        type: string
    required:
    - is_occupied
    - max_people
    - number
    - placement
    type: object
host: restoracio.fly.dev
info:
  contact: {}
  description: API for restaurant's management
  title: Restoracio
  version: "1.0"
paths:
  /admin:
    delete:
      consumes:
      - application/json
      description: creates a new admin
      operationId: admin-delete
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: AdminDelete
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: creates a new admin
      operationId: admin-create
      parameters:
      - description: Admin
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      summary: AdminCreate
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: creates a new admin
      operationId: admin-update
      parameters:
      - description: Admin
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Admin'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.apimodels.Message'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: AdminUpdate
      tags:
      - Admin
...
  /reserv:
    get:
      consumes:
      - application/json
      description: returns all reservations in a restaurant
      operationId: reserv-arst
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ReservDB'
            type: array
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: ReservationGetAllInRestaurant
      tags:
      - Reservation
    post:
      consumes:
      - application/json
      description: creates a new reservation
      operationId: reserv-create
      parameters:
      - description: Reservation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReservAPI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ReservDB'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: ReservationCreate
      tags:
      - Reservation
    put:
      consumes:
      - application/json
      description: updates a reservation
      operationId: reserv-update
      parameters:
      - description: Reservation
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ReservAPI'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: ReservationUpdate
      tags:
      - Reservation
  /reserv/{id}:
    delete:
      consumes:
      - application/json
      description: deletes a  reservation
      operationId: reserv-delete
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: ReservationDelete
      tags:
      - Reservation
    get:
      consumes:
      - application/json
      description: returns a reservation by id
      operationId: reserv-get-by-id
      parameters:
      - description: Reservation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ReservDB'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: ReservationGetByID
      tags:
      - Reservation
  /restaurant:
    delete:
      consumes:
      - application/json
      description: deletes a restaurant
      operationId: rest-delete
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: RestaurantDelete
      tags:
      - Restaurant
    get:
      consumes:
      - application/json
      description: returns an admin's restaurant
      operationId: rest-get-mine
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Restaurant'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: RestaurantGetMine
      tags:
      - Restaurant
    post:
      consumes:
      - application/json
      description: creates a new restaurant
      operationId: rest-create
      parameters:
      - description: Restaurant
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Restaurant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Restaurant'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: RestaurantCreate
      tags:
      - Restaurant
    put:
      consumes:
      - application/json
      description: updates a restaurant
      operationId: rest-update
      parameters:
      - description: Restaurant
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Restaurant'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: RestaurantUpdate
      tags:
      - Restaurant
  /restaurant/{id}:
    get:
      consumes:
      - application/json
      description: returns a restaurant by id
      operationId: rest-get-by-id
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Restaurant'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.apimodels.Message'
      security:
      - JWTAuth: []
      summary: RestaurantGetByID
      tags:
      - Restaurant
...
securityDefinitions:
  JWTAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"